+++ packages/inngest/src/components/execution/v1.ts
@@ -5,6 +5,7 @@ import { logPrefix } from "../../helpers/consts";
 import {
   ErrCode,
   deserializeError,
+  minifyPrettyError,
   prettyError,
   serializeError,
 } from "../../helpers/errors";
@@ -307,6 +308,7 @@ class V1InngestExecution extends InngestExecution implements IInngestExecution {
     await this.#state.hooks?.afterExecution?.();
 
     return newSteps.map<OutgoingOp>((step) => ({
+      displayName: step.displayName,
       op: step.op,
       id: step.hashedId,
       name: step.name,
@@ -314,12 +316,24 @@ class V1InngestExecution extends InngestExecution implements IInngestExecution {
     })) as [OutgoingOp, ...OutgoingOp[]];
   }
 
-  async #executeStep({ id, name, opts, fn }: FoundStep): Promise<OutgoingOp> {
+  async #executeStep({
+    id,
+    name,
+    opts,
+    fn,
+    displayName,
+  }: FoundStep): Promise<OutgoingOp> {
     this.#timeout?.clear();
     await this.#state.hooks?.afterMemoization?.();
     await this.#state.hooks?.beforeExecution?.();
 
-    const outgoingOp: OutgoingOp = { id, op: StepOpCode.RunStep, name, opts };
+    const outgoingOp: OutgoingOp = {
+      id,
+      op: StepOpCode.RunStep,
+      name,
+      opts,
+      displayName,
+    };
     this.#state.executingStep = outgoingOp;
     this.debug(`executing step "${id}"`);
 
@@ -446,7 +460,7 @@ class V1InngestExecution extends InngestExecution implements IInngestExecution {
         retriable = error.retryAfter;
       }
 
-      const serializedError = serializeError(error);
+      const serializedError = minifyPrettyError(serializeError(error));
 
       return { type: "function-rejected", error: serializedError, retriable };
     }